Ensure no overflow on addition for masking
The result of the addition should be uint64_t now because of
integral promotion.

+++ stempy/image.cpp
@@ -56,7 +56,9 @@ namespace {
     VTKM_EXEC_CONT
     OutputType operator()(const InputType& a, const InputType& b) const
     {
-      return OutputType(a[0] & a[1] + b[0] & b[1], a[0] & a[2] + b[0] & b[2]);
+      // Cast one of these to uint64_t to ensure no overflow on addition
+      return OutputType(static_cast<uint64_t>(a[0] & a[1]) + b[0] & b[1],
+                        static_cast<uint64_t>(a[0] & a[2]) + b[0] & b[2]);
     }
 
     VTKM_EXEC_CONT
